+++ TVS_Ventilation_DBX/TVS_TAP.h
@@ -255,11 +255,11 @@ static	TVS_TAP *add_new(
 	void checkRadiusTypeRound();
 	void checkRadiusTypeRect();
 	void checkTypeRoundTap();
-	Acad::ErrorStatus put_RadiusConst(int newVal);
+	Acad::ErrorStatus put_RadiusConst(double newVal);
 	int get_RadiusConst(void) const;
 	Acad::ErrorStatus put_RadiusTypeRect(int newVal);
 	int get_RadiusTypeRect(void) const;
-	Acad::ErrorStatus put_RadiusVariableParameter(int newVal);
+	Acad::ErrorStatus put_RadiusVariableParameter(double newVal);
 	int get_RadiusVariableParameter(void) const;
 	void setFlex(const bool &isFlex);
 protected:
